+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 
 ### Enhancements
 
-- Rename the `dsym` upload command to `xcode-build` to better reflect the command's purpose. [156](https://github.com/bugsnag/bugsnag-cli/pull/156)
-- Add the `xcode-archive` command to upload `.xcarchive` files from Xcode Archive builds to BugSnag [157](https://github.com/bugsnag/bugsnag-cli/pull/157)
+- Rename the `dsym` upload command to `xcode-build` to better reflect the command's purpose. `dsym` will be removed in the next major release [156](https://github.com/bugsnag/bugsnag-cli/pull/156)
 
 ## 2.7.0 (2024-11-26)